requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.2
